add audit search
