+++ audit/urls.py
@@ -1,0 +1,7 @@
+from . import views
+from django.urls import path
+
+app_name = 'audit'
+urlpatterns = [
+    path('search', views.AuditSearchView.as_view(), name='search'),
+]
