+++ audit/views.py
@@ -1,3 +1,12 @@
+from django.contrib.auth.mixins import LoginRequiredMixin
 from django.shortcuts import render
 
-# Create your views here.
+from audit.models import CofkUnionAuditLiteral
+from core.helper.view_utils import BasicSearchView, DefaultSearchView
+
+
+class AuditSearchView(LoginRequiredMixin, DefaultSearchView):
+
+    def get_queryset(self):
+        queryset = CofkUnionAuditLiteral.objects.all()
+        return queryset
