+++ core/templates/core/emlo_logined.html
@@ -29,7 +29,7 @@
                 </ul>
             </li>
             <li class="menu-item">
-                <a href="/audit/">Audit</a>
+                <a href="{% url 'audit:search' %}">Audit</a>
             </li>
             <li class="menu-item">
                 <a href="#!">{{ request.user.username }}</a>
