+++ siteedit2/settings/base.py
@@ -49,6 +49,7 @@ INSTALLED_APPS = [
     'manifestation',
     'person',
     'publication',
+    'audit',
 ]
 
 MIDDLEWARE = [
